Refactor dl to reusable DescriptionList component
